rename: 명시적으로 atom 이름 변경
